Move NamedFieldSelect into field-mobile component to fix storybook
