+++ src/component/field.js
@@ -1,10 +1,8 @@
 import React from 'react';
 import { Text as RNText, StyleSheet, View, ViewPropTypes } from 'react-native';
-import RNPickerSelect from 'react-native-picker-select';
 import PropTypes from 'prop-types';
 import { createStyles, maxWidth } from './media-query';
 import Text from './text';
-import ArrowDownIcon from '../asset/icon/arrow-down';
 import { TextInput, HorizontalExpandingTextInput } from './input';
 import { color, font, breakWidth } from './style';
 
@@ -99,52 +97,6 @@ export const NamedField = ({ name, children, style }) => (
 NamedField.propTypes = {
   name: PropTypes.string,
   children: PropTypes.oneOfType([PropTypes.string, PropTypes.array]),
-  style: ViewPropTypes.style,
-};
-
-//
-// Named Field Select
-//
-
-const namedSelectStyles = StyleSheet.create({
-  wrapper: {
-    flexDirection: 'row',
-    alignItems: 'center',
-  },
-});
-
-const pickerStyles = StyleSheet.create({
-  inputIOS: {
-    fontFamily: 'OpenSans Regular',
-    fontSize: font.sizeM,
-    lineHeight: font.lineHeightM + 3,
-    height: font.lineHeightM + 2 * 12,
-    color: color.blackText,
-    opacity: 0.75,
-  },
-  inputAndroid: {
-    fontFamily: 'OpenSans Regular',
-    fontSize: font.sizeM,
-    lineHeight: font.lineHeightM + 3,
-    height: font.lineHeightM + 2 * 12,
-    color: color.blackText,
-    opacity: 0.75,
-    padding: 0,
-  },
-});
-
-export const NamedFieldSelect = ({ name, style, ...props }) => (
-  <View style={[namedStyles.content, style]}>
-    <Text style={namedStyles.name}>{name}</Text>
-    <View style={namedSelectStyles.wrapper}>
-      <RNPickerSelect placeholder={{}} style={pickerStyles} {...props} />
-      <ArrowDownIcon height={22} width={22} stroke="#969596" />
-    </View>
-  </View>
-);
-
-NamedFieldSelect.propTypes = {
-  name: PropTypes.string,
   style: ViewPropTypes.style,
 };
