+++ src/view/pay-bitcoin-confirm-mobile.js
@@ -4,7 +4,8 @@ import { observer } from 'mobx-react';
 import PropTypes from 'prop-types';
 import Background from '../component/background';
 import MainContent from '../component/main-content';
-import { NamedField, NamedFieldSelect } from '../component/field';
+import { NamedField } from '../component/field';
+import { NamedFieldSelect } from '../component/field-mobile';
 import { Header, Title } from '../component/header';
 import { CancelButton, BackButton, SmallGlasButton } from '../component/button';
 import Card from '../component/card';
