+++ stories/component/field-story.js
@@ -4,7 +4,6 @@ import { action } from '@storybook/addon-actions';
 import {
   InputField,
   NamedField,
-  NamedFieldSelect,
   DetailField,
   AmountInputField,
 } from '../../src/component/field';
@@ -14,11 +13,6 @@ storiesOf('Field', module)
     <InputField placeholder="Input Field" onChangeText={action('input')} />
   ))
   .add('Named Field', () => <NamedField name="Label">Named field</NamedField>)
-  .add('Named Field Select', () => (
-    <NamedFieldSelect name="Label" onPress={action('select')}>
-      Named field Select
-    </NamedFieldSelect>
-  ))
   .add('Detail Field', () => (
     <DetailField name="Label">Detail field</DetailField>
   ))
